Moves `SsoTestSetup` to `servicemocks` package
